% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGroups.R
\name{randomGroups}
\alias{randomGroups}
\title{Creates a random list of partner groups}
\usage{
randomGroups(sheet, size = 2, seed = 8675309)
}
\arguments{
\item{sheet}{(required) is an object containing the list of student names.}

\item{size}{is the size of each group.}
}
\value{
\code{studentList} returns list of names and group numbers. .

\code{results} returns the original submissions (with equilibria and points per round added).

\code{grades} returns the aggregated points "won" by each student for the entire activity.
}
\description{
Creates a random list of partner groups
}
\details{
\code{partnerList} uses a random number to divide a \code{studentList} groups of size N.
}
